requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
